Remove warnings in pytest and old abstract class
