+++ cryojax/simulator/_scattering_theory/__init__.py
@@ -9,6 +9,5 @@ from .strong_phase_scattering_theory import (
     StrongPhaseScatteringTheory as StrongPhaseScatteringTheory,
 )
 from .weak_phase_scattering_theory import (
-    AbstractWeakPhaseScatteringTheory as AbstractWeakPhaseScatteringTheory,
     WeakPhaseScatteringTheory as WeakPhaseScatteringTheory,
 )
